Add files via upload
Conditional formatting is a robust feature in Power BI that you can leverage to help users gain insights. Typically, we see conditional formatting applied to the table and matrix visuals. For example, you might use a background colour, font colour, or icons for the values based on a set of rules. 
Conditional formatting of tables is valuable, but the capabilities of Power BI go beyond that.